+++ home-manager/default.nix
@@ -1,13 +1,7 @@
 {
   imports = [
     ../common/keyboard.nix
-    ./modules/git.nix
-    ./modules/gpg.nix
     ./modules/identity.nix
-    ./modules/pms.nix
-    ./modules/screen.nix
-    ./modules/tridactyl.nix
-    ./modules/zathura.nix
-    ./modules/zsh.nix
+    ./modules/programs
   ];
 }