+++ nixos/modules/features/default.nix
@@ -1,0 +1,7 @@
+{
+  imports = [
+    ./fonts.nix
+    ./shell.nix
+    ./utilities.nix
+  ];
+}