+++ nixos/modules/programs/default.nix
@@ -1,0 +1,9 @@
+{
+  imports = [
+    ./nix.nix
+    ./ranger.nix
+    ./tmux.nix
+    ./vim.nix
+    ./zsh.nix
+  ];
+}